restructured file organization
