+++ .eslintrc.json
@@ -60,9 +60,13 @@
         "map": [
           ["@assets", "./src/assets/"],
           ["@components", "./src/components/"],
+          ["@components", "./src/components"],
+          ["@ks", "./src/screens/ks"],
           ["@lib", "./src/lib/"],
+          ["@nsevp", "./src/screens/nsevp"],
           ["@root", "./src/"],
           ["@route", "./src/route/"],
+          ["@shared", "./src/screens/shared"],
           ["@styles", "./src/styles/"]
         ],
         "extensions": [".js", ".css"]
